Check link

+++ build/src/scripts/start-rtl.sh
@@ -2,7 +2,9 @@
 set -e
 
 # Create restore symlink
-ln -s /rtl/backup /rtl/backup/restore
+if [ ! -f /rtl/backup/restore ]; then
+  ln -s /rtl/backup /rtl/backup/restore
+fi
 
 # Configure settings from env vars
 envsubst < "source_RTL.conf" > "RTL.conf"
